+++ src/mi/dyn-info-list.c
@@ -30,5 +30,5 @@ HIDDEN unw_dyn_info_list_t _U_dyn_info_list;
 PROTECTED unw_word_t
 _U_dyn_info_list_addr (void)
 {
-  return (unw_word_t) &_U_dyn_info_list;
+  return (unw_word_t) (uintptr_t) &_U_dyn_info_list;
 }
