+++ src/setjmp/longjmp.c
@@ -63,7 +63,7 @@ _longjmp (jmp_buf env, int val)
       if (unw_set_reg (&c, UNW_REG_EH + 0, wp[JB_RP]) < 0
 	  || unw_set_reg (&c, UNW_REG_EH + 1, val) < 0
 	  || unw_set_reg (&c, UNW_REG_IP,
-			  (unw_word_t) &_UI_longjmp_cont))
+			  (unw_word_t) (uintptr_t) &_UI_longjmp_cont))
 	abort ();
 
       unw_resume (&c);
